% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalyZone.R
\name{anomalyZone}
\alias{anomalyZone}
\title{anomalyZone}
\usage{
anomalyZone(tree = NULL, outgroups = NULL, print.node = TRUE)
}
\arguments{
\item{tree}{phylogenetic tree read into R in phylo format}

\item{outgroups}{outgroups to root your tree}

\item{print.node}{prints the anomaly zone nodes if found}
}
\value{
a data.table of anomaly zone data calculated for all nodes in your tree
}
\description{
Applies the anomaly zone calculation across the entire tree (Degnan & Rosenberg 2006)
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
