% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAlignments.R
\name{filterAlignments}
\alias{filterAlignments}
\title{filterAlignments}
\usage{
filterAlignments(
  filter.summary = NULL,
  alignment.data = NULL,
  alignment.folder = NULL,
  format = c("folder", "concatenated"),
  min.alignments = 5,
  overwrite = TRUE
)
}
\arguments{
\item{filter.summary}{summary data file from filterSummary}

\item{alignment.data}{summary data file from alignmentSummary}

\item{alignment.folder}{folder of alignments to be filtered}

\item{format}{save format for alignments}

\item{min.alignments}{minimum number of alignments for a filtered set of data}

\item{overwrite}{overwrite if file exists?}
}
\value{
filters the alignments
}
\description{
Function for filtering alignments based on filter summary data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
