% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAnomalies.R
\name{filterAnomalies}
\alias{filterAnomalies}
\title{filterAnomalies}
\usage{
filterAnomalies(astral.directory = NULL, outgroups = NULL, filter.data = NULL)
}
\arguments{
\item{astral.directory}{directory of filtered astral results}

\item{outgroups}{outgroups to root your tree}

\item{filter.data}{your master filtered dataset summary stats}
}
\value{
a data.table of anomaly zone data calculated for all nodes in your tree
}
\description{
Applies the anomaly zone calculation across all filtered datasets and combines with filter summary data for downstream analyses
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
