% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterConcordance.R
\name{filterConcordance}
\alias{filterConcordance}
\title{filterConcordance}
\usage{
filterConcordance(
  input.dir = NULL,
  clade.list = NULL,
  outgroups = NULL,
  all.data = TRUE
)
}
\arguments{
\item{input.dir}{summary data file from filterSummary}

\item{clade.list}{a named list of clades of interest to test for concordance factors}

\item{outgroups}{outgroups to root the tree}

\item{all.data}{select T to obtain all the node data or F to obtain only nodes from clade list}
}
\value{
data.frame with concordance factor data for each filtered replicate
}
\description{
Function for obtaining filtered concordance factor data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
