% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeneTrees.R
\name{filterGeneTrees}
\alias{filterGeneTrees}
\title{filterGeneTrees}
\usage{
filterGeneTrees(
  filter.summary = NULL,
  alignment.data = NULL,
  genetree.folder = NULL,
  format = c("folder", "concatenated"),
  overwrite = TRUE,
  taxa.remove = NULL,
  min.trees = 5,
  min.n.samples = 4,
  min.sample.prop = NULL,
  make.polytomy = TRUE,
  polytomy.limit = 0,
  remove.node.labels = FALSE
)
}
\arguments{
\item{filter.summary}{summary data file from filterSummary}

\item{alignment.data}{summary data file from alignmentSummary}

\item{genetree.folder}{your target folder of gene trees that correspond to the alignments being filtered}

\item{format}{save format for genetrees}

\item{overwrite}{overwrite if file exists?}

\item{taxa.remove}{species that you would like removed from each gene tree}

\item{min.trees}{mimimum number of trees to keep filtered set}

\item{min.n.samples}{the minimum number of samples to keep a gene tree}

\item{min.sample.prop}{the minimum proportion of samples to keep a gene tree}

\item{polytomy.limit}{the value at which to collapse a node into a polytomy}

\item{remove.node.labels}{strips trees of node labels if downstream analyses give you trouble (not recommended)}

\item{collapse.polytomy}{collapses polytomies in the gene trees}
}
\value{
filters gene trees either in a folder or concatenated set of trees
}
\description{
Filters gene trees using provided filtration and alignment data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
