% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterStats.R
\name{filterStats}
\alias{filterStats}
\title{filterStats}
\usage{
filterStats(
  data = NULL,
  filter.name = c("alignment_length", "count_pis", "proportion_pis",
    "proportion_samples"),
  min.alignments = 5,
  filter.values = NULL,
  align.dataset = NULL
)
}
\arguments{
\item{data}{alignment summmary data from alignmentSummary function}

\item{filter.name}{Name of the filter being passed. Only one can be used at a time.}

\item{min.alignments}{minimum number of alignments to keep a filtration scheme}

\item{filter.values}{the filter values for the given scheme}

\item{align.dataset}{select three colors to plot your pie.plot}
}
\value{
Alignment statistics for the specific filteration scheme provided
}
\description{
Function for gather filtration statistics, used internally by filterSummary
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
