% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSummary.R
\name{filterSummary}
\alias{filterSummary}
\title{filterSummary}
\usage{
filterSummary(
  alignment.data = NULL,
  alignment.folder = NULL,
  dataset.name = NULL,
  file.out = NULL,
  length.filters = c(0),
  sample.filters = c(0),
  prop.pis.filters = c(0),
  count.pis.filters = c(0)
)
}
\arguments{
\item{alignment.data}{Alignment summary stats calculcated from summarizeAlignments}

\item{alignment.folder}{The alignment folder from which the stats were calculated from in alignment.data}

\item{dataset.name}{The name of your dataset, where all filtered datasets will be placed in this folder}

\item{file.out}{if you wish to save to file, provide a file name for the summary}

\item{length.filters}{Your selected length filters as a vector of values for the alignment length}

\item{sample.filters}{Your selected sampling fraction filters as a vector of values between 0-1}

\item{prop.pis.filters}{Your selected parsimony informatives sites filter as a vector of values between 0-1}

\item{count.pis.filters}{Your selected parsimony informatives sites filter as a vector of base pair counts}
}
\value{
Returns a summary of your selected filters applied to the dataset
}
\description{
Summarizes filtered datasets when provided with filter parameters
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
