% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.anomalyZone.R
\name{plot.anomalyZone}
\alias{plot.anomalyZone}
\title{plot.anomalyZone}
\usage{
\method{plot}{anomalyZone}(
  tree = NULL,
  data = NULL,
  outgroups = NULL,
  save.file = NULL,
  tip.label.size = 0.5,
  node.label.size = 1,
  edge.width = 3
)
}
\arguments{
\item{tree}{Tree that you wish to plot}

\item{data}{data.table from the anomalyZone function}

\item{outgroups}{provide outgroups for rooting plotted tree}

\item{save.file}{if you wish to save to file, put file name. Saves as PDF}

\item{tip.label.size}{size of the tip labels, passed to cex in plotting function}

\item{node.label.size}{size of the node label circles, passed to cex in plotting function}

\item{edge.width}{size of the branch edges, passed to edge.width in plot.phylo}
}
\value{
plots the phylogenetic tree and selected data associated with an AstralPlane object. Can optionally be saved to file as a PDF by giving save.file a file name.
}
\description{
Function for plotting data collected for the anomaly zone
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
