% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.filterCFAZ.R
\name{plot.filterCFAZ}
\alias{plot.filterCFAZ}
\title{plot.filterCFAZ}
\usage{
\method{plot}{filterCFAZ}(
  anomaly.zone.data = NULL,
  concordance.factors.data = NULL,
  save.plots = TRUE,
  output.dir = "Filter-Plots",
  focal.node = NULL,
  filter.name = c("alignment_length", "count_pis", "proportion_pis",
    "proportion_sample"),
  dataset.name = "all",
  plot.gcf = TRUE,
  plot.scf = TRUE,
  az.colors = c("#7BC143", "#DE3293"),
  m.shape = c(22, 21),
  min.trees = 10
)
}
\arguments{
\item{anomaly.zone.data}{table output from the filterAnomalies function}

\item{concordance.factors.data}{table output from the filterConcordance function}

\item{save.plots}{if you wish to save to file select TRUE}

\item{output.dir}{the name of the output directory to save the plots if TRUE above}

\item{focal.node}{the node annotated from the filterConcordance function}

\item{filter.name}{the filter to plot. Options include: alignment_length, count_pis, proportion_pis, proportion_sample}

\item{dataset.name}{from your main sets of data (i.e. exons, introns). All = plots of all them.}

\item{plot.gcf}{should the gene concordance factor be plotted?}

\item{plot.scf}{should the site concordance factor be plotted?}

\item{az.colors}{colors to indicate the anomaly zone. Default: Green: presence; Purple: absence}

\item{m.shape}{monophyly shape on the graph; circle = monophyletic; square paraphyletic}

\item{min.trees}{minimum number of trees to keep a filtration replicate. Default: 10}
}
\value{
plots a dot plot of each filter and the concordance factors for the focal node. Monophyly of that node is also shown.
}
\description{
Function for plotting data from concordance factors and the anomaly zone
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
